+++ src/NeustaPimcoreHttpCacheBundle.php
@@ -3,7 +3,7 @@
 namespace Neusta\Pimcore\HttpCacheBundle;
 
 use FOS\HttpCacheBundle\FOSHttpCacheBundle;
-use Neusta\Pimcore\HttpCacheBundle\DependencyInjection\CacheTagCollectorPass;
+use Neusta\Pimcore\HttpCacheBundle\DependencyInjection\CompilerPass\DisableCacheTagCollectorPass;
 use Pimcore\Extension\Bundle\AbstractPimcoreBundle;
 use Pimcore\Extension\Bundle\Traits\PackageVersionTrait;
 use Pimcore\HttpKernel\Bundle\DependentBundleInterface;
@@ -27,6 +27,6 @@ final class NeustaPimcoreHttpCacheBundle extends AbstractPimcoreBundle implement
     public function build(ContainerBuilder $container): void
     {
         parent::build($container);
-        $container->addCompilerPass(new CacheTagCollectorPass());
+        $container->addCompilerPass(new DisableCacheTagCollectorPass());
     }
 }
